+++ Samples/Disable Default PowerShell Session Config.ps1
@@ -1,0 +1,15 @@
+Configuration DisableDefaultPowerShell
+{
+    Import-DscResource -Module JeaDsc
+
+    JeaSessionConfiguration DnsManagementEndpoint
+    {
+        Name                = 'microsoft.powershell'
+        AccessMode          = 'Disabled'
+    }
+}
+
+Remove-Item -Path C:\DscTest\* -ErrorAction SilentlyContinue
+DisableDefaultPowerShell -OutputPath C:\DscTest -Verbose
+
+Start-DscConfiguration -Path C:\DscTest -Wait -Verbose -Force