+++ CHANGELOG.md
@@ -5,9 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Adding AccessMode to PSSessionConfiguration.
+
 ### Added
 
-- Adding AccessMode to PSSessionConfiguration.
 - Adding herited classes that contains helper methods.
 - Adding Reason class.
 - Adding Reasons property in JeaSessionConfiguration and JeaRoleCapabilities resources.